HACK: don't call setgroups when dropping privileges
The reason for this is that setgroups isn't supported in unprivileged
Linux user namepsaces. The reason for that is that setgroups can
actually allow you to *gain* privileges: If the permissions for a file
are set such that the group has less privileges than the world (for
example, 606) then dropping that group will increase our privileges.
This is more common in Linux systems with some variety of rich ACLs
enabled, but it's also done in some more conventional systems using
normal Unix permissions.

I could make a configuration option for this. But I sincerely doubt
OpenBSD will merge this...