HACK: allow lda to run as root
Not sure why this check is here, which means this removal is almost
certainly wrong. But, as a hack, this lets me deliver mail to, or as,
"root".

Concretely, since we're using lmtp, we don't need to (and should not)
setuid to the user that we're delivering mail to. We are running
everything as user namespace "root", so we add "user root" to the lmtp
delivery action so that we don't setuid to another user as part of
delivery.

But without this patch, delivery with the lmtp action fails at this
check. The check seems to designed to prevent delivering mail to root,
but it also stops our hack. We'll just remove the check for now...